+++ tests/TestCase.php
@@ -8,6 +8,8 @@ use PHPUnit\Framework\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
 {
+
+    public mixed $output = null;
     
     public function mock(string $method, $url)
     {