+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.1 - releases
+
+### Features
+ - Add prefix to routes
+
+### Fixes
+ - Fix `Too few arguments to function` for named parameters ([#2](https://github.com/AmolKumarGupta/php-router/issues/2))
+
+
 ## 0.1.0 - 2024-07-23
 
 ### Features
@@ -7,5 +16,3 @@
  - Add class based callback support
  - Add matchRoutes method
  - Add `get`, `post`, `put`, `delete` method for their corresponding HTTP method
-
-### Fixes